Initial commit.
Android manifest is being generated, including limited permissions. Not including access property yet. Most of the logic here should be in the simple android-permission.js file.
String localization is happening into a strings.xml file, into the correct strings-* directory.
Icons are being downloaded and placed in the correct place. This is fairly complete, except we're assuming that images are pngs.

Tests are non-meaningful, and there needs to be an optimist front end.