possible fix for Bug#1060473
Under still unknown conditions HttpFileLoader.copy fails to call
the callback while loading icons for the apk_project_builder. This
causes the build to hang and never decrement the activeBuilds counter
which eventually exhausts the concurrent build limit.

This change buffers the icon instead of streaming to make the error
handling more explicit.